Working on updating activity and onebooth to show correct text when user has no booths and no friends.
