+++ css/pageframe.css
@@ -4,11 +4,17 @@ body {
     text-align: center;
     padding: 0;
     margin: 0;
+    background-color: #515097;
+    font-family: 'Bitter', serif;
 }
 
 a {
     text-decoration: none;
     color: black;
+}
+
+.bootherLink, .boothLink {
+    font-weight: bold;
 }
 
 .loaderRegion {
@@ -244,7 +250,9 @@ a {
     }
 
     .sidebar_titleandbutton {
-        background-color: #FFF;
+        background-color: #87E1C1;
+        margin-top: 0.5rem;
+        border-radius: 0.5rem;
         color: #000;
     }
