+++ newbooth/templates/fileUpload.mst
@@ -1,0 +1,33 @@
+<center>
+    <a href = '{{baseUrl}}/uploadbooth'>
+        <div class = 'subheader' style = 'position: relative; width: 640px;' >
+            If you have problems with this page, click here to access the old upload page...
+        </div>
+    </a>
+    <div id = "previewspot" style = "width: 640px; height: 280px; border: 1px dashed black; box-shadow: 0 0 5px #AAAAAA;"></div>
+    <div style = "width: 80%;">
+    <form id = "uploadform" method = null>
+        <div id = "fileselectsection" style = "position: absolute; opacity: 0; z-index: -1; width: 0px; height: 0px;">
+        <input type="file" name="file" id="file" onChange = "showPreviewAuto(this.files, false)" />
+        </div>
+        <div id = "buttonspot">
+        </div>
+        <div class = "greensection">
+            <input type = "radio" name = "rotation" id = "0" value = 0 checked />
+            <label class = "selected" for="0"><h1>0&deg;</h1></label>
+            <input type = "radio" name = "rotation" id = "90"  value = 90 />
+            <label for="90"><h1>90&deg;</h1></label>
+            <input type = "radio" name = "rotation" id = "180"  value = 180 />
+            <label for="180"><h1>180&deg;</h1></label>
+            <input type = "radio" name = "rotation" id = "270"  value = 270 />
+            <label for="270"><h1>270&deg;</h1></label>
+        </div>
+        <textarea style='width: 640px; height: 200px;' id = 'blurb' name='blurb' placeholder='Write a blurb to go with this picture'></textarea><br />
+        <div class = 'bigbutton' id = 'submit_button'>Upload Booth</div>
+        <progress id = "progress" max="100" value="0" style = "width: 100%">
+        </progress>
+    </form>
+    <div id = 'status' style = 'height: 32px;'></div>
+    </div>
+</center>
+<script type = "text/javascript" src = "{{baseUrl}}/newbooth/file.js"></script>
